Studio: Tidying and consolidating.
Merge CameraAnimationManager.ts into Studio.ts. Move a few global functions into the only classes that were calling them. Re-arrange some declarations for tidiness. Remove unneeded exports.
