+++ src/Studio.ts
@@ -1,16 +1,16 @@
 import * as Viewer from './viewer';
 import { UI, Checkbox, setElementHighlighted, createDOMFromString } from './ui';
 import { FloatingPanel } from './DebugFloaters';
-import { CameraAnimationManager, InterpolationStep } from './CameraAnimationManager';
 import { StudioCameraController } from './Camera';
-import { clamp, computeEulerAngleRotationFromSRTMatrix, getMatrixAxisZ, Vec3UnitY, Vec3Zero } from './MathHelpers';
+import { clamp, computeEulerAngleRotationFromSRTMatrix, getMatrixAxisZ, invlerp, Vec3UnitY, Vec3Zero } from './MathHelpers';
 import { mat4, ReadonlyMat4, vec3 } from 'gl-matrix';
 import { GlobalSaveManager } from './SaveManager';
+import { getPointHermite } from './Spline';
 
-export const MILLISECONDS_IN_SECOND = 1000.0;
 export const CLAPBOARD_ICON = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 64" height="20" fill="white"><path d="M61,22H14.51l3.41-.72h0l7.74-1.64,2-.43h0l6.85-1.46h0l1.17-.25,8.61-1.83h0l.78-.17,9-1.91h0l.4-.08L60,12.33a1,1,0,0,0,.77-1.19L59.3,4.3a1,1,0,0,0-1.19-.77l-19,4-1.56.33h0L28.91,9.74,27.79,10h0l-9.11,1.94-.67.14h0L3.34,15.17a1,1,0,0,0-.77,1.19L4,23.11V60a1,1,0,0,0,1,1H61a1,1,0,0,0,1-1V23A1,1,0,0,0,61,22ZM57,5.8l.65.6.89,4.19-1.45.31L52.6,6.75ZM47.27,7.88,51.8,12,47.36,13,42.82,8.83ZM37.48,10,42,14.11l-4.44.94L33,10.91ZM27.7,12l4.53,4.15-4.44.94L23.26,13Zm-9.78,2.08,4.53,4.15L18,19.21l-4.53-4.15ZM19.49,29H14.94l3.57-5h4.54Zm9-5h4.54l-3.57,5H24.94ZM39,45.88l-11,6A1,1,0,0,1,26.5,51V39A1,1,0,0,1,28,38.12l11,6a1,1,0,0,1,0,1.76ZM39.49,29H34.94l3.57-5h4.54Zm10,0H44.94l3.57-5h4.54ZM60,29H54.94l3.57-5H60Z"/></svg>`;
 const UNDO_ICON = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" height="16"><g transform="translate(0,-952.36218)"><path overflow="visible" style="" d="m 39.999997,975.36218 -31.9999995,25.00002 31.9999995,25 0,-14 c 1.7024,-0.08 31.3771,-0.033 52.000005,18 -8.252999,-25.4273 -34.173805,-35.48722 -52.000005,-40.00002 z" fill="#ffffff" stroke="none"/></g></svg>`;
 const REDO_ICON = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" height="16"><g transform="translate(0,-952.36218)"><path d="m 60,975.36216 32,24.99994 -32,25.0001 0,-12.0001 c -1.7024,-0.08 -31.3771,-2.0334 -52,16.0001 8.253,-25.4274 34.1738,-37.48724 52,-42.00004 z" style="" overflow="visible" fill="#ffffff" stroke="none"/></g></svg>`;
+const MILLISECONDS_IN_SECOND = 1000.0;
 const MIN_ANIMATION_LENGTH_SEC = 1;
 const MAX_ANIMATION_LENGTH_SEC = 300;
 
@@ -22,7 +22,35 @@ export interface Keyframe {
     useAutoTangent: boolean;
 }
 
-export class KeyframeTrack {
+export class InterpolationStep {
+    public pos: vec3 = vec3.create();
+    public lookAtPos: vec3 = vec3.create();
+    public bank: number = 0;
+}
+
+export interface CameraAnimation {
+    posXTrack: KeyframeTrack;
+    posYTrack: KeyframeTrack;
+    posZTrack: KeyframeTrack;
+    lookAtXTrack: KeyframeTrack;
+    lookAtYTrack: KeyframeTrack;
+    lookAtZTrack: KeyframeTrack;
+    bankTrack: KeyframeTrack;
+    loop: boolean;
+}
+
+const enum KeyframeTrackType {
+    posXTrack    = 0b0000001,
+    posYTrack    = 0b0000010,
+    posZTrack    = 0b0000100,
+    lookAtXTrack = 0b0001000,
+    lookAtYTrack = 0b0010000,
+    lookAtZTrack = 0b0100000,
+    bankTrack    = 0b1000000,
+    allTracks    = 0b1111111,
+}
+
+class KeyframeTrack {
     public keyframes: Keyframe[] = [];
 
     public addKeyframe(kf: Keyframe) {
@@ -88,7 +116,7 @@ export class KeyframeTrack {
         const val = (k2.value - k0.value) * 0.5;
         const prevDuration = k1.time - k0.time;
         const nextDuration = k2.time - k1.time;
-        k1.tangentIn  = val * (2 * prevDuration) / (prevDuration + nextDuration);
+        k1.tangentIn = val * (2 * prevDuration) / (prevDuration + nextDuration);
         k1.tangentOut = val * (2 * nextDuration) / (prevDuration + nextDuration);
     }
 
@@ -113,47 +141,6 @@ export class KeyframeTrack {
         this.keyframes.sort((a, b) => a.time - b.time);
     }
 
-}
-
-export interface CameraAnimation {
-    posXTrack: KeyframeTrack;
-    posYTrack: KeyframeTrack;
-    posZTrack: KeyframeTrack;
-    lookAtXTrack: KeyframeTrack;
-    lookAtYTrack: KeyframeTrack;
-    lookAtZTrack: KeyframeTrack;
-    bankTrack: KeyframeTrack;
-    loop: boolean;
-}
-
-export const enum KeyframeTrackType {
-    posXTrack    = 0b0000001,
-    posYTrack    = 0b0000010,
-    posZTrack    = 0b0000100,
-    lookAtXTrack = 0b0001000,
-    lookAtYTrack = 0b0010000,
-    lookAtZTrack = 0b0100000,
-    bankTrack    = 0b1000000,
-    allTracks    = 0b1111111,
-}
-
-function getTrackByType(animation: CameraAnimation, trackType: KeyframeTrackType): KeyframeTrack {
-    if (trackType === KeyframeTrackType.posXTrack)
-        return animation.posXTrack;
-    else if (trackType === KeyframeTrackType.posYTrack)
-        return animation.posYTrack;
-    else if (trackType === KeyframeTrackType.posZTrack)
-        return animation.posZTrack;
-    else if (trackType === KeyframeTrackType.lookAtXTrack)
-        return animation.lookAtXTrack;
-    else if (trackType === KeyframeTrackType.lookAtYTrack)
-        return animation.lookAtYTrack;
-    else if (trackType === KeyframeTrackType.lookAtZTrack)
-        return animation.lookAtZTrack;
-    else if (trackType === KeyframeTrackType.bankTrack)
-        return animation.bankTrack;
-    else
-        throw "whoops";
 }
 
 /**
@@ -613,6 +600,98 @@ class Timeline {
                 return true;
         }
         return false;
+    }
+}
+
+export class CameraAnimationManager {
+    // The animation to play back.
+    private animation: Readonly<CameraAnimation>;
+    private elapsedTimeMs: number;
+
+    // The animation's duration.
+    public durationMs: number;
+
+
+    private findKeyframe(frames: Readonly<Keyframe[]>, time: number): number {
+        for (let i = 0; i < frames.length; i++)
+            if (time < frames[i].time)
+                return i;
+        return -1;
+    }
+
+    private calcTrackDuration(track: Readonly<KeyframeTrack>): number {
+        // Assume it's sorted.
+        if (track.keyframes.length > 0)
+            return track.keyframes[track.keyframes.length - 1].time;
+        else
+            return 0;
+    }
+
+    private calcAnimationDuration(animation: Readonly<CameraAnimation>): number {
+        let duration = 0;
+        duration = Math.max(duration, this.calcTrackDuration(animation.posXTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.posYTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.posZTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.lookAtXTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.lookAtYTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.lookAtZTrack));
+        duration = Math.max(duration, this.calcTrackDuration(animation.bankTrack));
+        return duration;
+    }
+
+    private getCurrentTrackValue(track: KeyframeTrack, time: number): number {
+        const idx1 = this.findKeyframe(track.keyframes, time);
+        if (idx1 === 0)
+            return track.keyframes[0].value;
+        if (idx1 < 0)
+            return track.keyframes[track.keyframes.length - 1].value;
+
+        const idx0 = idx1 - 1;
+        const k0 = track.keyframes[idx0], k1 = track.keyframes[idx1];
+
+        const t = invlerp(k0.time, k1.time, time);
+        return getPointHermite(k0.value, k1.value, k0.tangentOut, k1.tangentIn, t);
+    }
+
+    private calcAnimationPose(dst: InterpolationStep, animation: Readonly<CameraAnimation>, time: number): void {
+        const posX = this.getCurrentTrackValue(animation.posXTrack, time);
+        const posY = this.getCurrentTrackValue(animation.posYTrack, time);
+        const posZ = this.getCurrentTrackValue(animation.posZTrack, time);
+        const lookAtX = this.getCurrentTrackValue(animation.lookAtXTrack, time);
+        const lookAtY = this.getCurrentTrackValue(animation.lookAtYTrack, time);
+        const lookAtZ = this.getCurrentTrackValue(animation.lookAtZTrack, time);
+        vec3.set(dst.pos, posX, posY, posZ);
+        vec3.set(dst.lookAtPos, lookAtX, lookAtY, lookAtZ);
+        dst.bank = this.getCurrentTrackValue(animation.bankTrack, time);
+    }
+
+    public initAnimationPlayback(animation: Readonly<CameraAnimation>, startTimeMs: number) {
+        this.animation = animation;
+        this.elapsedTimeMs = startTimeMs;
+        this.durationMs = this.calcAnimationDuration(animation);
+    }
+
+    public setElapsedTime(t: number): void {
+        this.elapsedTimeMs = t;
+
+        if (this.animation.loop)
+            this.elapsedTimeMs = this.elapsedTimeMs % this.durationMs;
+    }
+
+    public updateElapsedTime(dt: number): void {
+        this.setElapsedTime(this.elapsedTimeMs + dt);
+    }
+
+    public getAnimFrame(outInterpStep: InterpolationStep, time: number = this.elapsedTimeMs) {
+        this.calcAnimationPose(outInterpStep, this.animation, time);
+    }
+
+    public isAnimationFinished(): boolean {
+        return !this.animation.loop && this.elapsedTimeMs >= this.durationMs;
+    }
+
+    public getElapsedTimeSeconds(): number {
+        return this.elapsedTimeMs / MILLISECONDS_IN_SECOND;
     }
 }
 
@@ -1254,7 +1333,7 @@ export class StudioPanel extends FloatingPanel {
         this.timelineElementsCanvas.addEventListener('keyframeDeselected', (e: Event) => { this.hideKeyframeControls(); });
         this.timelineElementsCanvas.addEventListener('keyframeIconMovedEvent', (e: Event) => {
             this.timeline.selectedKeyframeIcon!.keyframesMap.forEach((v, trackType) => {
-                getTrackByType(this.animation, trackType).reSort();
+                this.getTrackByType(this.animation, trackType).reSort();
             });
         });
 
@@ -1485,13 +1564,32 @@ export class StudioPanel extends FloatingPanel {
         }
     }
 
+    private getTrackByType(animation: CameraAnimation, trackType: KeyframeTrackType): KeyframeTrack {
+        if (trackType === KeyframeTrackType.posXTrack)
+            return animation.posXTrack;
+        else if (trackType === KeyframeTrackType.posYTrack)
+            return animation.posYTrack;
+        else if (trackType === KeyframeTrackType.posZTrack)
+            return animation.posZTrack;
+        else if (trackType === KeyframeTrackType.lookAtXTrack)
+            return animation.lookAtXTrack;
+        else if (trackType === KeyframeTrackType.lookAtYTrack)
+            return animation.lookAtYTrack;
+        else if (trackType === KeyframeTrackType.lookAtZTrack)
+            return animation.lookAtZTrack;
+        else if (trackType === KeyframeTrackType.bankTrack)
+            return animation.bankTrack;
+        else
+            throw "whoops";
+    }
+
     private onChangeTangentInput(input: HTMLInputElement): void {
         if (this.timeline.selectedKeyframeIcon && input.value) {
             const val = parseFloat(input.value);
             if (!Number.isNaN(val)) {
                 const trackType = parseInt(input.dataset.track!, 10);
                 const kf = this.timeline.selectedKeyframeIcon.keyframesMap.get(trackType)!;
-                getTrackByType(this.animation, trackType).setCustomTangent(kf, val);
+                this.getTrackByType(this.animation, trackType).setCustomTangent(kf, val);
                 this.updatePreviewSteps();
             }
         }
@@ -1813,7 +1911,7 @@ export class StudioPanel extends FloatingPanel {
         const kfMap = new Map<KeyframeTrackType, Keyframe>();
 
         const addLoopKeyframe = (trackType: KeyframeTrackType): void => {
-            const track = getTrackByType(this.animation, trackType);
+            const track = this.getTrackByType(this.animation, trackType);
             const loopKeyframe = makeLoopKeyframe(track);
             track.addKeyframe(loopKeyframe);
             kfMap.set(trackType, loopKeyframe);
